[feature] Completed functional Schedule UI for Timesheets and Breaks. Though the UI may require some cosmestic changes/tweaks over time.
The UI is built on Backbone.js and is fully interactive, and supports in
place editing of any Break Entry or Time Entry.

Syncing the model
operations to the rails backend is also supported, so we can have
instant save (during an edit).

Note: this is just the frontend backbone that is complete, we still
have remaining backend saving (but that's relatively quick)